user profile image done

+++ Util/deleteFile.js
@@ -2,22 +2,26 @@ const fs = require('fs');
 
 const deleteSingleFile = (filePath) => {
     if (filePath) {
-        fs.unlink(filePath, (err) => {
-            if (err) {
-                throw (err);
-            }
-        })
+        if (fs.existsSync(filePath)) {
+            fs.unlink(filePath, (err) => {
+                if (err) {
+                    throw (err);
+                }
+            })
+        }
     }
 }
 
 const deleteMultiFile = (filePath) => {
     filePath.map(path => {
         if (path) {
-            fs.unlink(path, (err) => {
-                if (err) {
-                    throw (err);
-                }
-            })
+            if (fs.existsSync(path)) {
+                fs.unlink(path, (err) => {
+                    if (err) {
+                        throw (err);
+                    }
+                });
+            }
         }
     })
 }
